Update Employee.java
Explanation of the process: 
In the main method, Employee (10245,"Bunga") stored in obj1 calls the constructor Employee(int id, String name). In this method the value 10245 is stored in empId and Bunga is stored in empName. 

Next, in the main method, Employee(92232,"Joko") stored in obj2 calls the constructor Employe(int id, String name). In this method the value 92232 is stored in empId and Joko is stored in empName. 

Then, return to the main method, obj1.info() which calls the info() method which will print an output of the value of obj1